Started to factor out code common for more organisations
